// src/openapi/models/LogParserResponse_4dfe1dd_LogEntry.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type LogParserResponse_4dfe1dd_LogEntry = {
    timestamp: string;
    unit: number;
    subunit: number;
    unit_subunit_id: number;
    ini_filename: string;
    code: string;
    description: string;
    value: string;
    type_um: string;
    snapshot: string;
    color: string;
};